# backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Vereinsfinder API
  version: 1.0.0
  description: REST API für Vereine, Angebote, Eigenschaften, Suche und Import
servers:
  - url: http://localhost:8080
paths:
  /api/vereine:
    get:
      summary: Liste aller Vereine
      responses:
        '200':
          description: Erfolgreich
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Verein'
    post:
      summary: Verein anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Verein'
      responses:
        '201':
          description: Erstellt
  /api/vereine/{id}:
    get:
      summary: Verein nach ID abrufen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Verein'
        '404':
          description: Nicht gefunden
    put:
      summary: Verein aktualisieren
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Verein'
      responses:
        '200':
          description: Erfolgreich
    delete:
      summary: Verein löschen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Gelöscht
...
  /api/import/vereine:
    post:
      summary: Importiert Vereine aus einer JSON-Datei
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import abgeschlossen
...
components:
  schemas:
    Verein:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        beschreibung:
          type: string
        websiteUrl:
          type: string
        adresse:
          type: string
        telefonnummer:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        angebote:
          type: array
          items:
            $ref: '#/components/schemas/Angebot'
        eigenschaften:
          type: array
          items:
            $ref: '#/components/schemas/Eigenschaft'
    Angebot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        beschreibung:
          type: string
        vereine:
          type: array
          items:
            $ref: '#/components/schemas/Verein'
        eigenschaften:
          type: array
          items:
            $ref: '#/components/schemas/Eigenschaft'
    Eigenschaft:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        beschreibung:
          type: string
        vereine:
          type: array
          items:
            $ref: '#/components/schemas/Verein'
        angebote:
          type: array
          items:
            $ref: '#/components/schemas/Angebot'
...